Change object to Stateful object to show issue with states
Now you can check the checkboxes next to the items, but they will not follow their items as list is being reordered. This is because State is not directly paired with Widgets, but their Element objects, that in turn only change reference to now different Widget.
(just see the video)

udemy: https://www.udemy.com/course/learn-flutter-dart-to-build-ios-android-apps/learn/lecture/37143888#overview